fix: folder structure to the given format
