+++ src/providers/ReduxProvider.tsx
@@ -2,7 +2,7 @@
 
 import { Provider } from "react-redux";
 import { ReactNode } from "react";
-import { store } from "../lib/redux/utils/store";
+import { store } from "../../lib/redux/utils/store";
 
 export default function Providers({ children }: { children: ReactNode }) {
   return <Provider store={store}>{children}</Provider>;
